Move all functionalities of plots to io/plots
